move NoMatchMatch to match.rb
