+++ lib/fluent/agent.rb
@@ -17,6 +17,7 @@
 require 'fluent/configurable'
 require 'fluent/plugin'
 require 'fluent/output'
+require 'fluent/match'
 
 module Fluent
   #
@@ -157,36 +158,6 @@ module Fluent
     end
 
     def handle_emits_error(tag, es, error)
-    end
-
-    class NoMatchMatch
-      def initialize(log)
-        @log = log
-        @count = 0
-      end
-
-      def emit_events(tag, es)
-        # TODO use time instead of num of records
-        c = (@count += 1)
-        if c < 512
-          if Math.log(c) / Math.log(2) % 1.0 == 0
-            @log.warn "no patterns matched", tag: tag
-            return
-          end
-        else
-          if c % 512 == 0
-            @log.warn "no patterns matched", tag: tag
-            return
-          end
-        end
-        @log.on_trace { @log.trace "no patterns matched", tag: tag }
-      end
-
-      def start
-      end
-
-      def shutdown
-      end
     end
   end
 end
